Samsung msm7x27: install init, ueventd files from common config.
Previously we needed to push the init and ueventd files from each
per-device config. With this change it is no longer necessary, letting
us streamline future changes to init files.

Note: this commit also installs the new bluez init.rc, and adds a template
parts.rc to satisfy build. Per-device configs can override the parts.rc to
add custom services, but it can also be useful if we create a shared
"SamsungParts" helper app in the future.

Change-Id: I4ad14ad85095b3a16e35d98feef4b11a9e91c3df